refactor: 소셜 로그인 관련 파일 구조 변경 및 service명 변경
