+++ src/main/java/com/Alchive/backend/service/UserService.java
@@ -1,21 +1,17 @@
 package com.Alchive.backend.service;
 
 import com.Alchive.backend.config.Code;
-import com.Alchive.backend.config.ErrorCode;
 import com.Alchive.backend.config.GeneralException;
 import com.Alchive.backend.domain.User;
 import com.Alchive.backend.dto.request.UserUpdateRequest;
 import com.Alchive.backend.repository.UserRepository;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
-
-import java.util.NoSuchElementException;
 
 @RequiredArgsConstructor
 @Service
-public class UserDetailService {
+public class UserService {
     private final UserRepository userRepository;
 
     public User getUserDetail(Long userId) {
